+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Improvements
 
 - Integrated GW2 API data for profession and elite specialization IDs to prevent unnecessary event logging
+- Better skill detection, auto attack chaisn will not skip skills afterwards
 
 ## 0.2.0 - 2025-10-17