+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added more advanced Skill Icon coloring Logic
 - Don't show Cascading Corruption
 - Show Rusing Justice only once per cast
+- Added debug info for the mumble context
 
 ## 0.3.0 - 2025-10-19