+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Skill info on icon hover in horizontal mode
 - Only draw in training area
+- Auto reload if ooc
 
 ### Improvements